remove temporary flags as minimum_send, minimum_recv, asym_power, local_win_size and global_win_size revert back to int based calculations
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@985 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ schedule.c
@@ -150,8 +150,7 @@ void schedule_forward_packet(struct orig_node *orig_node, struct bat_packet *in,
 		}
 
 		/* apply hop penalty */
-		((struct bat_packet *)forw_node_new->pack_buff)->tq = ((float) (((struct bat_packet *)forw_node_new->pack_buff)->tq)) * ((float)( TQ_MAX_VALUE - hop_penalty)/ (TQ_MAX_VALUE));
-
+		((struct bat_packet *)forw_node_new->pack_buff)->tq = (((struct bat_packet *)forw_node_new->pack_buff)->tq * (TQ_MAX_VALUE - hop_penalty)) / (TQ_MAX_VALUE);
 		debug_output( 4, "forwarding: tq_orig: %i, tq_avg: %i, tq_forw: %i, ttl_orig: %i, ttl_forw: %i \n", in->tq, tq_avg, ((struct bat_packet *)forw_node_new->pack_buff)->tq, in->ttl - 1, ((struct bat_packet *)forw_node_new->pack_buff)->ttl );
 
 		forw_node_new->send_time = get_time_msec();
